requirements.txt
amqp==5.1.1
attrs==22.1.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
exceptiongroup==1.0.4
Flask==2.2.2
flower==1.0.0
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
importlib-metadata==4.8.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.6
requests==2.28.1
six==1.16.0
tomli==2.0.1
tornado==6.2
typing-extensions==4.4.0
urllib3==1.26.13
validators==0.20.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
redis==4.4.0
